feat: display: Add support for the CH1115 display
As it turns out: The CH1115 is just a tiny bit more advanced version to
the SH11106. They are so similar in fact, that it is possible to drive
the CH1115 using an SH11106 driver. The additional features of the
CH1115 are mostly in regards to power saving. It additionally also
supports vertical and horizontal text scrolling and a "breathing
effect".

As such the support just needed a few tweaks in the board DTS file. So
here we are.

As the display is a bit larger than the display area reserved for it on
the canon calc two additional properties to setup the offset/dimensions
of the display have been introduced.

Together with the dts, there's an example application which demonstrates
the features of the display.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>