feat: driver: Add skeleton GPIO input kscan driver
The goal of this driver is to develop a key matrix scanning driver which
accepts pairs of GPIO (input/output) pins and is able to trigger the
input subsystem if they are triggered.

So here we go, with a skeleton build of that which already integrates
with DTS, the build system and Kconfig.

One would use this driver for example as:

  #include <zephyr/dt-bindings/input/input-event-codes.h>

  / {
      keyboard {
          compatible = "gpio-kscan";

          key_0 {
              input-gpios = <&gpio0 13 (GPIO_PULL_UP | GPIO_ACTIVE_LOW)>;
              output-gpios = <&gpio0 13 GPIO_ACTIVE_LOW>;
              zephyr,code = <INPUT_KEY_0>;
          };
      };
  };

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>