feat: Add skeletal os directory
The purpose of the os directory is to contain important code, that needs
to run in order for the application to use the given hardware easily and
without much trouble.

This is of course already mostly served by zephyr, but not everything
that can help, is too small / too specific or generally not fit as a
driver. To contain those would be the purpose of the os code base.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>