feat: gpio_kscan: Functioning polling driver
This iteration of the `gpio_kscan` driver is fully capable of receiving
keys from a key-matrix (as would be specified in the overlay file).

As there's no clear assortment of input/output pins (and output pins
need to be triggered), it uses a polling based system instead of
interrupts to read the matrix.

Events are passed to the input subsystem.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>