feat: Add bad-apple application
Plays bad apple on the calculator (or any other compatible video) using
heatshrink compression and dumping the video on the internal flash.

Sadly the current Zephyr version doesn't support octal SPI RAM so LZ4
could not be used, as it requires a bit more space than is available on
internal RAM (when using 64kb block size).

By pressing a key on the keyboard, the playback will reset.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>