feat: cancalc; Move important Kconfigs to shield
There are some config parameters that are always used for the canon
calculator shield like display, i2c, …. As I was previously just playing
around these were set by applications, which of course isn't needed
anymore now as we got it into a working state. So assume it for every
build target by setting them in the shield.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>