feat: shield: Add LVGL support
This patch adds a new UI framework besides the character framebuffer
called LVGL. This is known to be a fully featured UI framework for
embedded devices, which for us has the benefit of non-monospaced,
unicode and UTF-8 font rendering.

One can now select in the shield configuration through a simple choice
field which UI framework one wants to use.

Also moved the display dts node from user defined to chosen nodes, as it
has a reserved field for this.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>