feat: os: Add base object that represents current screen
In LVGL one cannot virtually limit screen sizes, but instead they
inherit their properties directly from the display they are associated
with.

The recommended solution here is to instead create a base object on the
screen and add all elements there, as positional properties are in
reference to the parent.

Such a base object can now be accessed from the OS using `get_screen()`.

fix:
  I additionally needed to set the initialisation priority far lower, as
  the `os_display_init` method would otherwise run before LVGL was able
  to start. Using a priority of 99 now to do initialisation just before
  the main.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>