fix: random crashes while using LVGL / stack overflows
The log thread did not have a large enough stack size to fulfil its
duties, so in a last attempt it took over the threads by others
resulting in those nice and very hard to debug stack overflows.

I was able to reverse the problem by using a combination of a few Zephyr
flags / features:

- CONFIG_STACK_CANARIES=y
  - With it software would crash at random spots due to stack canaries
    being triggered. If using it, note that the thread triggering the
    canary, isn't the one who destroyed it.
- CONFIG_THREAD_ANALZYER
  - a handy tool by Zephyr too see where the thread is at its limit.
    It gave me the idea that the log thread was responsible as it showed
    a >97% stack usage for that thread.

If thread names are not enabled one can use the memory map
(`build/zephyr/zephyr.map`) to see which symbol / region is reserved for
the given address.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>